Cleaned up most code.
Cleaned up most code. Also fixed that the tubename.php file only gets
the tubename when a channel ID is given (otherwise an error will occur).
The subcount.php file will now convert an username to channel ID in
order to get subscribers when an username is given instead of showing 0
subscribers.